Feature/tests

* Add reports tests
* Add basic trips tests
* Finish covering by tests of trips on 100%, merge add/edit templates into one
* Add tests on meals page
* Add auth tests
* Remove dangerous methods from developer page, add tests
* Add CLI commands tests

+++ organizer/users.py
@@ -1,4 +1,4 @@
-from flask import Blueprint, render_template, request, redirect, url_for
+from flask import Blueprint, render_template, request, redirect, url_for, flash, abort, g
 from werkzeug.security import generate_password_hash
 
 from organizer.auth import login_required_group
@@ -19,15 +19,32 @@ def index():
 @bp.route('/add', methods=['GET', 'POST'])
 @login_required_group(AccessGroup.Administrator)
 def add():
+    groups = [x.name for x in AccessGroup]
+    add_page = 'users/add.html'
+
     if request.method == 'POST':
-        # check values here
         name = request.form['name']
+        if not name:
+            flash('Empty name provided')
+            return render_template(add_page, groups=groups)
 
         password = request.form['password']
+        if not password:
+            flash('Empty password provided')
+            return render_template(add_page, groups=groups)
+
         pass_hash = generate_password_hash(password)
 
         group = request.form['group']
-        access_group = AccessGroup[group]
+        if not group:
+            flash('Empty group provided')
+            return render_template(add_page, groups=groups)
+
+        try:
+            access_group = AccessGroup[group]
+        except KeyError:
+            flash('Wrong group provided')
+            return render_template(add_page, groups=groups)
 
         with get_session() as session:
             session.add(User(name=name, password=pass_hash,
@@ -36,13 +53,21 @@ def add():
 
         return redirect(url_for('users.index'))
 
-    return render_template('users/add.html', groups=[x.name for x in AccessGroup])
+    return render_template(add_page, groups=groups)
 
 
 @bp.route('/remove/<int:user_id>')
 @login_required_group(AccessGroup.Administrator)
 def delete(user_id):
     with get_session() as session:
+        if user_id == g.user.id:
+            # it is not possible to remove current user
+            abort(403)
+
+        user = session.query(User).filter(User.id == user_id).first()
+        if not user:
+            abort(404)
+
         session.query(User).filter(User.id == user_id).delete()
         session.commit()
         return redirect(url_for('users.index'))
